add option to copy all at once

+++ apigateway/routes/misc.py
@@ -126,7 +126,7 @@ def handle_data_migration():
     mongo_collection = mongo_database['twoMinuteStats']
     mongo_collection.update_many(query, {'$set': {'userId': user_id}})
 
-    return {'message': 'success'}, 202
+    return {'message': 'success'}, 200
 
 
 @app.route('/app_logs', methods=['POST'])
@@ -169,22 +169,41 @@ def handle_test_data_copy(principal_id=None):
     stats_collection_test = mongo_database['athleteStatsTest']
     # completed_exercises_collection_test = mongo_database['completedExercisesTest']
 
-    # clear all existing data
-    readiness_collection.delete_many({"user_id": user_id})
-    plan_collection.delete_many({"user_id": user_id})
-    completed_exercises_collection.delete_many({"athlete_id": user_id})
-    stats_collection.delete_one({"athlete_id": user_id})
-    # get data from test collections
-    rs_datastore = DailyReadinessDatastore()
-    readiness_surveys = rs_datastore.get(user_id,
-                                         last_only=False,
-                                         mongo_collection=readiness_collection_test)
-    daily_plan_datastore = DailyPlanDatastore()
-    daily_plans = daily_plan_datastore.get(user_id,
-                                           mongo_collection=plan_collection_test)
-    athlete_stats_datastore = AthleteStatsDatastore()
-    athlete_stats = athlete_stats_datastore.get(user_id,
-                                                mongo_collection=stats_collection_test)
+
+    if "copy_all" in request.json and request.json["copy_all"]:
+        athlete_stats_datastore = AthleteStatsDatastore()
+        mongo_results = stats_collection_test.find()
+        athlete_stats = []
+        for mongo_result in mongo_results:
+            athlete_stats.append(athlete_stats_datastore.get_athlete_stats_from_mongo(mongo_result))
+        all_users = [stat.athlete_id for stat in athlete_stats]
+        # get data from test collections
+        rs_datastore = DailyReadinessDatastore()
+        readiness_surveys = rs_datastore.get(all_users,
+                                             last_only=False,
+                                             mongo_collection=readiness_collection_test)
+        daily_plan_datastore = DailyPlanDatastore()
+        daily_plans = daily_plan_datastore.get(all_users,
+                                               mongo_collection=plan_collection_test)
+    else:
+        # clear all existing data
+        readiness_collection.delete_many({"user_id": user_id})
+        plan_collection.delete_many({"user_id": user_id})
+        completed_exercises_collection.delete_many({"athlete_id": user_id})
+        stats_collection.delete_one({"athlete_id": user_id})
+        # get data from test collections
+        rs_datastore = DailyReadinessDatastore()
+        readiness_surveys = rs_datastore.get(user_id,
+                                             last_only=False,
+                                             mongo_collection=readiness_collection_test)
+        if len(readiness_surveys) == 0:
+            raise ForbiddenException("User not present in personas")
+        daily_plan_datastore = DailyPlanDatastore()
+        daily_plans = daily_plan_datastore.get(user_id,
+                                               mongo_collection=plan_collection_test)
+        athlete_stats_datastore = AthleteStatsDatastore()
+        athlete_stats = [athlete_stats_datastore.get(user_id,
+                                                     mongo_collection=stats_collection_test)]
 
     update_dates(readiness_surveys, daily_plans, athlete_stats, event_date)
     rs_datastore.put(readiness_surveys)
@@ -194,13 +213,14 @@ def handle_test_data_copy(principal_id=None):
     return {'message': 'success'}, 202
 
 def update_dates(rs_surveys, daily_plans, athlete_stats, event_date):
-    athlete_today = parse_date(athlete_stats.event_date)
+    athlete_today = parse_date(athlete_stats[0].event_date)
     day_diff = (event_date - athlete_today).days
     for rs_survey in rs_surveys:
         rs_survey.event_date += datetime.timedelta(days=day_diff)
 
     for plan in daily_plans:
         plan.event_date = format_date(parse_date(plan.event_date) + datetime.timedelta(days=day_diff))
-    athlete_stats.event_date = format_date(event_date)
+    for stat in athlete_stats:
+        stat.event_date = format_date(event_date)
